fix(421): break runtime->codegen import cycle; correct State docs + counts
Addresses the PR #789 review (#421 decomposition).

Import cycle (pr-test-analyzer): _validate_wrap_handle was parked in the compile-time vera/codegen/memory.py, but it is a runtime heap concern (called only by heap.py's _wrap_handle).  That created vera/runtime/heap.py -> vera.codegen.memory -> codegen/__init__ -> api.py -> vera.runtime.decimal -> back to a partially-initialised heap for _WRAP_KIND_DECIMAL, so a cold 'import vera.runtime.<x>' (no prior vera.codegen import) raised ImportError -- undercutting the very modularity the decomposition adds.  Move _validate_wrap_handle to heap.py (its real home; no other runtime->codegen edge remains), repoint its 6 unit tests, and add TestRuntimePackageImportHygiene421 -- a fresh-interpreter cold-import of all 15 runtime submodules (RED before the move, green after).

State docs accuracy (comment-analyzer): State is the one family registered UNCONDITIONALLY (it loops over result.state_types) and it threads state_store back into ExecuteResult.state -- so the 'registered conditionally, owns its own state, feeds nothing back' generalisation was wrong.  Corrected the api.py IO-section comment, the vera/README.md 'family module' paragraph, and the state.py docstring (which had claimed state_store is 'created here and internal') to say: most families are stateless/conditional; the two stateful ones (Decimal, State) thread a single store passed as one parameter that execute() reads back; IO differs in DEGREE (seven interwoven pieces).

Line counts (comment-analyzer): the IO comment grew api.py to 1,182 (not 1,165); the move makes memory.py 67 and heap.py 1,021; runtime/ is 3,540.  Fixed in README + CHANGELOG.  TESTING.md/ROADMAP.md test totals bumped for the new test (4759->4760, test_runtime_traps 66->67).
