+++ vera/README.md
@@ -122,8 +122,8 @@ execute(compile_result, ...)    # → run WASM via wasmtime
 | `  extensions.py` | 146 | | vera/speculativeEdit proof-delta | |
 | `  server.py` | 169 | | pygls wiring, single-session serialisation | |
 | `codegen/` | 6,618 | Compile | Codegen orchestrator (mixin package) | `compile()`, `execute()` |
-| `  api.py` | 1,165 | | Public API, dataclasses, `compile()`/`execute()` orchestration, core IO host bindings (#421) | |
-| `  memory.py` | 124 | | Compile-time ADT layout helpers (`ConstructorLayout`, alignment) (#421) | |
+| `  api.py` | 1,182 | | Public API, dataclasses, `compile()`/`execute()` orchestration, core IO host bindings (#421) | |
+| `  memory.py` | 67 | | Compile-time ADT layout helpers (`ConstructorLayout`, alignment) (#421) | |
 | `  core.py` | 711 | | CodeGenerator class, orchestration, ability op rewriting (Pass 1.6) | |
 | `  modules.py` | 392 | | Cross-module registration + call detection (C7e) | |
 | `  registration.py` | 258 | | Pass 1 forward declarations, ADT layout | |
@@ -133,9 +133,9 @@ execute(compile_result, ...)    # → run WASM via wasmtime
 | `  contracts.py` | 282 | | Runtime pre/postconditions, old state snapshots | |
 | `  assembly.py` | 856 | | WAT module assembly, `$alloc`, `$gc_collect` | |
 | `  compilability.py` | 310 | | Compilability checks, state handler scanning | |
-| `runtime/` | 3,470 | Execute | wasmtime host layer (#421): traps + per-effect host-binding families | `register_*()`, `WasmTrapError` |
+| `runtime/` | 3,540 | Execute | wasmtime host layer (#421): traps + per-effect host-binding families | `register_*()`, `WasmTrapError` |
 | `  traps.py` | 475 | | `WasmTrapError`, `_classify_trap`, source-backtrace resolution | |
-| `  heap.py` | 967 | | WASM memory marshalling primitives, ADT/Option/Array/bucket codecs, `_ShadowGuard`, shared collection helpers | |
+| `  heap.py` | 1,021 | | WASM memory marshalling primitives, ADT/Option/Array/bucket codecs, `_ShadowGuard`, shared collection helpers | |
 | `  collections.py` | 16 | | `_VAL_WASM_TYPES` value-type dispatch table (shared by Map/Set) | |
 | `  <effect>.py` ×12 | 2,006 | | one `register_<effect>(linker, …)` per family: random, math, md, json, regex, html, map, set, decimal, http, inference, state | |
 | `tester.py` | 750 | Test | Z3-guided input generation, WASM execution, tier classification | `test()` |
@@ -532,7 +532,7 @@ The two-pass architecture mirrors the type checker: pass 1 registers all functio
 
 Before #421, `execute()` and every effect's host bindings lived in one ~4,358-line `codegen/api.py`. The wasmtime host layer is now factored into `vera/runtime/`: trap classification (`traps.py`), WASM memory marshalling (`heap.py`, `collections.py`), and **one module per optional effect family**, each exposing a single `register_<family>(linker, …)` that defines and registers its host callbacks. `execute()` calls these in sequence instead of inlining ~3,000 lines of branches. The compiled `.wasm` import interface is unchanged — this is an internal refactor, not a contract change.
 
-**What counts as a "family" module.** Each of the twelve (`random`, `math`, `md`, `json`, `regex`, `html`, `map`, `set`, `decimal`, `http`, `inference`, `state`) is a *pluggable adapter*: registered conditionally (`if result.<effect>_ops_used`), owning its own state or none, and feeding nothing back into `execute()`'s return value. Map/Set/Decimal marshal opaque handles; Json/Html/Markdown/Regex bridge Python parsers; Http/Inference wrap network calls; Random/Math/State are thin shims. Any shared state a family needs is passed explicitly as a parameter (e.g. `register_decimal(linker, ops, decimal_store, host_store_refs)`); `heap.py` holds the marshalling primitives they all call.
+**What counts as a "family" module.** Each of the twelve (`random`, `math`, `md`, `json`, `regex`, `html`, `map`, `set`, `decimal`, `http`, `inference`, `state`) is a *pluggable adapter* with minimal coupling to `execute()`. Map/Set/Decimal marshal opaque handles; Json/Html/Markdown/Regex bridge Python parsers; Http/Inference wrap network calls; Random/Math/State are thin shims. Most are stateless and registered conditionally (`if result.<effect>_ops_used`). The two stateful ones — Decimal and State — keep a single Python-side store that `execute()` creates and reads back (Decimal's handle store feeds the GC decref hook and `host_store_sizes`; State's cell stacks feed `ExecuteResult.state`), passed as one explicit parameter (e.g. `register_decimal(linker, ops, decimal_store, host_store_refs)`), keeping each family a clean unit. `heap.py` holds the marshalling primitives they all call.
 
 **Why IO is *not* a family module.** IO stays inline in `execute()` by design. Unlike the twelve adapters, IO is execute()'s **observation channel**: its host callbacks write into state that *becomes the return value* — `output_buf`/`stderr_buf` → `ExecuteResult.stdout`/`stderr`, `last_violation` → the trap diagnostic via `_classify_trap`, `tee_stdout` → the live-streaming decision — and it shares the `_VeraExit` Ctrl-C exception with execute()'s exit handling. Extracting the twelve adapters *reduced* coupling (each became self-contained); extracting IO would not — it would relocate a naturally cohesive unit across a file boundary behind a 7-field context object that both the host callbacks and the result-building code would thread through. Cohesion of a genuinely-coupled unit outweighs uniform "every effect lives in `runtime/`" placement. (By Vera's *surface* model IO is an effect like any other; by the *compiler's* internal structure it is execute()'s I/O substrate. The decomposition follows the compiler's structure — the principle that each module be a cohesive, independently-testable unit.)
