+++ vera/codegen/memory.py
@@ -1,8 +1,7 @@
 """ADT memory-layout utilities for the Vera code generator.
 
-Pure data/math helpers -- constructor memory layout, WASM value-type
-sizes and alignment, and the #578 wrap-handle range guard.  No wasmtime
-or runtime dependency, so they import cleanly into the compiler
+Pure data/math helpers -- constructor memory layout, and WASM value-type sizes and
+alignment.  No wasmtime or runtime dependency, so they import cleanly into the compiler
 (`vera/codegen/*`, `vera/wasm/*`) and the runtime alike.  Extracted from
 `api.py` (#421); re-exported from `vera.codegen.api` for back-compat.
 """
@@ -39,61 +38,6 @@ class ConstructorLayout:
                 f"nat_fields (len {len(self.nat_fields)}) must match "
                 f"field_offsets (len {len(self.field_offsets)}) or be empty"
             )
-
-
-def _validate_wrap_handle(
-    raw_handle: object, kind: int, body_ptr: int,
-) -> None:
-    """#578 invariant: raw_handle must fit in 31 unsigned bits.
-
-    Wrapper ADTs store ``raw_handle | 0x80000000`` at body offset 4 so
-    the in-heap field is structurally outside the conservative-scan
-    heap-range check (`heap_ptr` is hard-capped at 0x80000000 by the
-    `$alloc` heap-ceiling guard).  The unwrap site recovers the raw
-    handle with ``& 0x7FFFFFFF``.  Both directions break silently
-    outside ``[0, 0x80000000)``:
-
-    - Negative ints have bit 31 set in two's complement.
-      ``raw_handle | 0x80000000`` is a no-op and the unwrap mask
-      returns the WRONG handle.
-    - Values ``>= 0x80000000`` alias into the top half and collide
-      with the tag-bit pattern.
-    - Values ``>= 0x100000000`` truncate on ``_write_i32`` and
-      silently lose information.
-    - Non-int values would ``TypeError`` deeper in the stack;
-      catching them here makes the diagnostic actionable.
-    - ``bool`` values: Python's ``bool`` subclasses ``int``, so
-      ``isinstance(True, int)`` is ``True`` and the value would
-      pass an ``isinstance``-only check, silently aliasing to
-      handles 1 and 0.  The strict ``type(raw_handle) is int``
-      check rejects bools without accepting any int subclass.
-
-    Practical alloc counters are bounded well below 2^31 — a 2B-handle
-    session is wall-clock infeasible — but a silent round-trip
-    failure is exactly the corruption class #578 sought to eliminate.
-    Fail fast.
-
-    Module-level helper so it can be unit-tested directly without
-    standing up a wasmtime instance (``_wrap_handle`` is nested
-    inside ``execute()`` and not importable on its own).
-    """
-    # ``type(x) is int`` rather than ``isinstance(x, int)`` so we
-    # reject ``bool`` (which would otherwise pass — bool subclasses
-    # int in Python and silently aliases to handles 0 / 1).
-    if not (
-        type(raw_handle) is int
-        and 0 <= raw_handle < 0x80000000
-    ):
-        raise RuntimeError(
-            f"#578: raw_handle={raw_handle!r} (kind={kind!r}, "
-            f"body_ptr={body_ptr!r}) is outside the valid "
-            f"range [0, 0x80000000); cannot tag for the "
-            f"conservative-scan disjointness invariant.  "
-            f"Host-store handle counters must be unsigned "
-            f"31-bit integers.  Either a counter overflowed, "
-            f"a negative sentinel flowed in, or a non-integer "
-            f"value flowed into _wrap_handle."
-        )
 
 
 def _wasm_type_size(wt: str) -> int:
